Consolidate pre-operation PKGBUILD review and warnings into unified `PkgbuildReviewDialog`
- Removed legacy `PackageBuildDialog` and `PkgbuildWarningDialog`.
- Introduced `PkgbuildDiff` model for unified diff generation and presentation.
- Added `PkgbuildReviewDialog` to replace separate dialogs with a single review UI.
- Updated services and flows to utilize `PkgbuildReviewDialog` during operations.
- Added unit tests for `PkgbuildDiff` calculation and correctness.
