+++ Shelly.Gtk/Services/GenericQuestionService.cs
@@ -5,7 +5,6 @@ namespace Shelly.Gtk.Services;
 public class GenericQuestionService : IGenericQuestionService
 {
     public event EventHandler<GenericQuestionEventArgs>? Question;
-    public event EventHandler<PackageBuildEventArgs>? PackageBuildRequested;
     public event EventHandler<GenericDialogEventArgs>? Dialog;
     
     public event EventHandler<ToastMessageEventArgs>? ToastMessageRequested;
@@ -13,11 +12,6 @@ public class GenericQuestionService : IGenericQuestionService
     public void RaiseQuestion(GenericQuestionEventArgs args)
     {
         Question?.Invoke(this, args);
-    }
-
-    public void RaisePackageBuild(PackageBuildEventArgs args)
-    {
-        PackageBuildRequested?.Invoke(this, args);
     }
 
     public void RaiseDialog(GenericDialogEventArgs args)
