+++ Shelly.Gtk/Windows/AUR/AurUpdate.cs
@@ -390,25 +390,9 @@ public class AurUpdate(
             {
                 lockoutService.Show(T("Installing..."));
 
-                var packageBuilds = await privilegedOperationService.GetAurPackageBuild(selectedPackages);
-
-                if (packageBuilds.Count != 0)
-                {
-                    foreach (var pkgbuild in packageBuilds)
-                    {
-                        if (pkgbuild.PkgBuild == null) continue;
-
-                        var buildArgs = new PackageBuildEventArgs(T("Displaying Package Build {0}", pkgbuild.Name),
-                            pkgbuild.PkgBuild);
-                        genericQuestionService.RaisePackageBuild(buildArgs);
-
-                        if (!await buildArgs.ResponseTask)
-                        {
-                            return;
-                        }
-                    }
-                }
-
+                // The PKGBUILD review/diff is now surfaced through the unified
+                // wire-based PkgbuildReviewDialog during the operation, so the
+                // legacy pre-operation PKGBUILD prompt is no longer raised here.
                 var result =
                     await privilegedOperationService.UpdateAurPackagesAsync(selectedPackages,
                         _runChecksCheck.GetActive());
